feat: refine event repository implementation
