+++ apps/auth-api/tsconfig.app.json
@@ -8,5 +8,5 @@
     "target": "es2015"
   },
   "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"],
-  "include": ["src/**/*.ts", "../../libs/shared/src/config.ts"]
+  "include": ["src/**/*.ts"]
 }
