+++ libs/shared/src/domains/plan-subscriptions/repositories/database/implementation/prisma-plan-subscriptions-database.repository.ts
@@ -1,8 +1,8 @@
 // users.repository.ts
 import { ConflictException, Injectable } from '@nestjs/common';
 import { PrismaService } from '../../../../../infra/storage/prisma/prisma.service';
-import { CreatePlanSubscriptionDto } from '../../../../dto/create-plan-subscription.dto';
-import { PlanSubscription } from '../../../../entities/plan-subscription.entity';
+import { CreatePlanSubscriptionDto } from '../../../dto/create-plan-subscription.dto';
+import { PlanSubscription } from '../../../entities/plan-subscription.entity';
 import { PlanSubscriptionsDatabaseRepository } from '../plan-subscriptions-database.repository';
 
 @Injectable()
