+++ libs/shared/src/domains/plan-subscriptions/repositories/database/plan-subscriptions-database.repository.ts
@@ -1,7 +1,7 @@
 import { Logger } from '@nestjs/common';
 import { featureFlags } from '../../../../config';
-import { CreatePlanSubscriptionDto } from '../../../dto/create-plan-subscription.dto';
-import { PlanSubscription } from '../../../entities/plan-subscription.entity';
+import { CreatePlanSubscriptionDto } from '../../dto/create-plan-subscription.dto';
+import { PlanSubscription } from '../../entities/plan-subscription.entity';
 import { InMemoryPlanSubscriptionsDatabaseRepository } from './implementation/in-memory-plan-subscriptions-database.repository';
 import { PrismaPlanSubscriptionsDatabaseRepository } from './implementation/prisma-plan-subscriptions-database.repository';
